requirements.txt
langchain
PyPDF2==3.0.1
pypdf
python-dotenv
streamlit
altair
tiktoken
huggingface-hub
unstructured
sentence_transformers
pinecone-client
ctransformers
